# openapi/agent-api.yaml
openapi: 3.1.0
info:
  title: Busibox Agent Server API
  version: 1.0.0
  description: |
    Production-grade agent server with Pydantic AI for executing AI agents with tool calls,
    dynamic agent/workflow management, and OAuth2 token forwarding to Busibox services.
    
    ## Authentication
    All endpoints (except `/health`) require Bearer token authentication with Busibox JWT.
    
    JWT tokens are issued by the AuthZ service (`authz-lxc:8010`) using OAuth2 token exchange.
    This service validates JWTs using the AuthZ JWKS endpoint (`/.well-known/jwks.json`).
    See `openapi/authz-api.yaml` for token issuance details.
    
    ## Features
    - Dynamic agent definition management with personal/built-in agents
    - Tool and workflow CRUD operations
    - Agent run execution with SSE streaming
    - Intelligent query routing via dispatcher
    - Performance evaluation with scorers
    - Schedule management for cron-based runs
    
  contact:
    name: Busibox Infrastructure Team
    email: infra@busibox.local
servers:
  - url: http://agent-lxc:8000
    description: Production (agent-lxc container)
  - url: http://10.96.200.30:8000
    description: Test environment
  - url: http://localhost:8000
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /agents/tools:
    get:
      summary: List active tools
      description: Returns all active tool definitions
      tags: [agents, tools]
      responses:
        '200':
          description: List of tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolDefinitionRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    post:
      summary: Create tool definition
      description: Create a new tool definition
      tags: [agents, tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolDefinitionCreate'
      responses:
        '201':
          description: Tool created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolDefinitionRead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /agents/tools/{tool_id}:
    get:
      summary: Get tool by ID
      description: Get individual tool definition by ID
      tags: [tools]
      parameters:
        - $ref: '#/components/parameters/ToolId'
      responses:
        '200':
          description: Tool definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolDefinitionRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      summary: Update tool definition
      description: |
        Update a custom tool definition.
        Built-in tools cannot be modified (returns 403).
        Version number increments automatically.
      tags: [tools]
      parameters:
        - $ref: '#/components/parameters/ToolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolDefinitionUpdate'
      responses:
        '200':
          description: Tool updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolDefinitionRead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Cannot modify built-in tools
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      summary: Delete tool definition
      description: |
        Soft-delete a custom tool.
        Built-in tools cannot be deleted.
        Tools in use by active agents cannot be deleted (returns 409).
      tags: [tools]
      parameters:
        - $ref: '#/components/parameters/ToolId'
      responses:
        '204':
          description: Tool deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Cannot delete built-in tools
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Tool is in use by active agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: resource_in_use
                  message:
                    type: string
                    example: Tool is in use by active agents
                  agents:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        name:
                          type: string
  /agents/workflows:
    get:
      summary: List active workflows
      description: Returns all active workflow definitions
      tags: [agents, workflows]
      responses:
        '200':
          description: List of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowDefinitionRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    post:
      summary: Create workflow definition
      description: |
        Create a new workflow definition with step validation.
        Workflows are created with created_by set to authenticated user.
      tags: [agents, workflows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowDefinitionCreate'
      responses:
        '201':
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDefinitionRead'
        '400':
          description: Invalid workflow definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /agents/workflows/{workflow_id}:
    get:
      summary: Get workflow by ID
      description: Get individual workflow definition by ID
      tags: [workflows]
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: Workflow definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDefinitionRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      summary: Update workflow definition
      description: |
        Update a workflow definition.
        Validates workflow steps before saving.
        Version number increments automatically.
      tags: [workflows]
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowDefinitionUpdate'
      responses:
        '200':
          description: Workflow updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDefinitionRead'
        '400':
          description: Invalid workflow steps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      summary: Delete workflow definition
      description: |
        Soft-delete a workflow.
        Workflows with active scheduled runs cannot be deleted.
      tags: [workflows]
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '204':
          description: Workflow deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Workflow has active scheduled runs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Busibox JWT token
  parameters:
...
    ToolId:
      name: tool_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Tool UUID
    WorkflowId:
      name: workflow_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Workflow UUID
...
  schemas:
...
    ToolDefinitionCreate:
      type: object
      properties:
        name:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          maxLength: 120
        description:
          type: string
          maxLength: 1000
        parameters:
          type: object
          description: JSON schema for tool parameters
        implementation:
          type: string
          description: Python code or reference to implementation
        is_active:
          type: boolean
          default: true
      required:
        - name
        - description
    ToolDefinitionRead:
      allOf:
        - $ref: '#/components/schemas/ToolDefinitionCreate'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            version:
              type: integer
            is_builtin:
              type: boolean
            created_by:
              type: string
              nullable: true
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
          required:
            - id
            - version
            - is_builtin
            - created_at
            - updated_at
    ToolDefinitionUpdate:
      type: object
      properties:
        description:
          type: string
          maxLength: 1000
        parameters:
          type: object
        implementation:
          type: string
        is_active:
          type: boolean
    WorkflowDefinitionCreate:
      type: object
      properties:
        name:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          maxLength: 120
        description:
          type: string
          maxLength: 1000
        steps:
          type: array
          items:
            type: object
            properties:
              agent_id:
                type: string
                format: uuid
              input_mapping:
                type: object
              output_key:
                type: string
        is_active:
          type: boolean
          default: true
      required:
        - name
        - steps
    WorkflowDefinitionRead:
      allOf:
        - $ref: '#/components/schemas/WorkflowDefinitionCreate'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            version:
              type: integer
            created_by:
              type: string
              nullable: true
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
          required:
            - id
            - version
            - created_at
            - updated_at
    WorkflowDefinitionUpdate:
      type: object
      properties:
        description:
          type: string
          maxLength: 1000
        steps:
          type: array
          items:
            type: object
        is_active:
          type: boolean
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
      required:
        - detail
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...